chahua: 重构 debug_panel.js —— 拆出纯工具与渲染树两个模块
debug_panel.js 重构后是前端最大文件（972 行）。按「一模块一职责」拆：

- debug_helpers.js（NEW）—— 常量 + 格式化 + 小 DOM 构件 + 工具源识别。
  无状态、无 import 的纯叶子，被 render / panel 共用。
- debug_render.js（NEW）—— renderTurn 及其子函数（renderCandidates /
  renderMessage / renderTool …）。纯函数：只吃 turnRecord、不碰闭包状态，
  renderTurn 是唯一对外出口。
- debug_panel.js —— 留实时 + 历史的有状态装配（turns / turnNodes / 历史
  索引共享同一套 evict）。972 → 579 行。

实时 turn 与历史索引共享 turns/turnNodes/evictOldest 是真实耦合，故有状态
状态机整体留在 debug_panel.js 不强拆 —— 只抽出零共享状态的纯部分。

承重契约不动：evict + swapRowBackToIndex 还原索引行、fromHistory 分支、
visibility gate、capture_prompts=false 的 null 区分均原样保留。

三个文件均过 node --check；前端无自动化测试，建议合并前在 Electron 里
手动冒烟（开调试抽屉 / 实时 turn / 点历史索引行展开详情 / 折叠）。
