+++ app/renderer/debug_panel.js
@@ -24,115 +24,22 @@
 // 内容换成详情卡片（详情进 ``turns`` Map，与实时 turn 共用 ``MAX_TURNS_IN_MEMORY``
 // 上限）。详情被 evict → 还原成轻量索引行（``<li>`` 节点不删，保留点击复拉入口，
 // docs §6.1 §10 评审第 1+3 条承重墙）。
+//
+// turn 渲染树（``renderTurn`` 及其子函数）见 ./debug_render.js；纯工具（格式化 /
+// 小 DOM 构件 / 工具源识别）见 ./debug_helpers.js。本文件留实时 + 历史的有状态
+// 装配（``turns`` / ``turnNodes`` / 历史索引共享同一套 evict）。
 
 import { EventType, Inbound, ScoreKind, Status } from "./events.js";
-
-// keep in sync with chahua/task_tools.py::TASK_WRITE_ARTIFACT_TOOL_NAME
-const TASK_WRITE_ARTIFACT = "task_write_artifact";
+import { renderTurn } from "./debug_render.js";
+import {
+  formatTs,
+  scoringPathLabel,
+  formatHistoryWinners,
+  deriveArtifactPath,
+  classifyToolSource,
+} from "./debug_helpers.js";
 
 const MAX_TURNS_IN_MEMORY = 50;
-
-// scoring_path 徽标文案 —— scoring / mention / broadcast 沿用既有英文标识；
-// handoff_* 这类对用户不直观的 wire 值给中文短标。data-path 仍存原值供 CSS 配色。
-const SCORING_PATH_LABELS = Object.freeze({
-  handoff_delegate: "指派",
-  handoff_review: "请审",
-  handoff_panel: "圆桌",
-});
-
-// handoff turn 顶部提示条整句。与 SCORING_PATH_LABELS（折叠徽标短标）分开维护：
-// delegate / review 的短标恰好是动词、panel 的「圆桌」是名词，提示条要完整动词
-// 短语，不能统一从短标拼。
-const HANDOFF_NOTE_TEXT = Object.freeze({
-  handoff_delegate: "由用户指派",
-  handoff_review: "由用户请审",
-  handoff_panel: "由用户发起圆桌",
-});
-function scoringPathLabel(path) {
-  return SCORING_PATH_LABELS[path] || path;
-}
-
-// handoff turn 的判定信号 —— scoring_path 以 ``handoff_`` 起头（live turn_start 与
-// 历史 jsonl 都带这个字段，比 trigger.kind 在两条路径上都可用）。
-function isHandoffPath(path) {
-  return typeof path === "string" && path.startsWith("handoff_");
-}
-
-// 跟 .gitignore / docs §不变量同口径：仅 task_write_artifact 派生 artifact 路径。
-// 后续若加新写盘工具按 tool 名扩；不挂 ArtifactDetector。
-//
-// name 校验镜像 ``tasks_store._validate_artifact_name``（``/`` / ``\`` / ``..`` /
-// 前缀 ``.``）—— 非法名会被写盘层拒、不会落盘，前端记一条"幽灵 artifact"会误导
-// 用户。后端 ``transport_bridge._maybe_record_artifact_path`` 同口径过滤。
-function deriveArtifactPath(toolName, args, taskId) {
-  if (toolName !== TASK_WRITE_ARTIFACT) return null;
-  if (!taskId) return null;
-  if (!args || typeof args !== "object") return null;
-  const name = args.name;
-  if (typeof name !== "string" || !name) return null;
-  if (name.includes("/") || name.includes("\\") || name.includes("..")) return null;
-  if (name.startsWith(".")) return null;
-  return `tasks/${taskId}/artifacts/${name}`;
-}
-
-function formatTs(ms) {
-  if (typeof ms !== "number" || !Number.isFinite(ms)) return "";
-  const d = new Date(ms);
-  const pad = (n) => String(n).padStart(2, "0");
-  return `${pad(d.getHours())}:${pad(d.getMinutes())}:${pad(d.getSeconds())}`;
-}
-
-function formatScore(score) {
-  if (typeof score !== "number" || !Number.isFinite(score)) return "—";
-  return score.toFixed(2);
-}
-
-function makeBadge(text, className = "") {
-  const el = document.createElement("span");
-  el.className = className;
-  el.textContent = text;
-  return el;
-}
-
-function makePre(text) {
-  const pre = document.createElement("pre");
-  pre.className = "debug-pre";
-  pre.textContent = text;
-  return pre;
-}
-
-// 空 winners + n_messages=0 = "全员低分的空 turn"，给个明显的视觉提示，避免与
-// "winner 数据丢失" 混淆。
-function formatHistoryWinners(entry) {
-  const winners = Array.isArray(entry.winners) ? entry.winners : [];
-  if (winners.length > 0) return winners.join(", ");
-  if (entry.n_messages === 0) return "（无人接话）";
-  return "—";
-}
-
-function makeDisabledPromptHint() {
-  const el = document.createElement("div");
-  el.className = "debug-pre-disabled";
-  el.textContent = "prompt 捕获已关（room.toml `[debug] capture_prompts = true` 重启后可见）";
-  return el;
-}
-
-// "label + body" 是 candidate / message 渲染里出现 5 次的模板：一个小标签 + 一个
-// pre/disabled 体块。集中在这里，渲染端调一行。
-function appendLabeledPre(parent, label, text) {
-  const lab = document.createElement("div");
-  lab.className = "debug-pre-label";
-  lab.textContent = label;
-  parent.appendChild(lab);
-  parent.appendChild(makePre(text));
-}
-function appendLabeledDisabled(parent, label) {
-  const lab = document.createElement("div");
-  lab.className = "debug-pre-label";
-  lab.textContent = label;
-  parent.appendChild(lab);
-  parent.appendChild(makeDisabledPromptHint());
-}
 
 // ── 内部数据形态 ─────────────────────────────────────────────────────
 //
@@ -427,19 +334,6 @@ export function createDebugPanel({ panelEl, bodyEl, clearBtnEl, sendInbound }) {
     rerenderTurn(turnId);
   }
 
-  // MCP 源识别 —— 与后端 debug_recorder._classify_tool_source 同算法（前端复刻，
-  // 避免再开个 envelope 字段透传）。
-  function classifyToolSource(tool) {
-    if (typeof tool !== "string" || !tool) return { source: "unknown", mcp_server: null };
-    if (tool.startsWith("mcp__")) {
-      const rest = tool.slice("mcp__".length);
-      const idx = rest.indexOf("__");
-      const server = idx > 0 ? rest.slice(0, idx) : rest;
-      return { source: "mcp", mcp_server: server || null };
-    }
-    return { source: "builtin", mcp_server: null };
-  }
-
   // ── 历史 turn 索引 ──────────────────────────────────────────────────
 
   function applyTurnsIndex(index) {
@@ -652,7 +546,7 @@ export function createDebugPanel({ panelEl, bodyEl, clearBtnEl, sendInbound }) {
     return record;
   }
 
-  // ── 渲染 ─────────────────────────────────────────────────────────────
+  // ── 重渲 ─────────────────────────────────────────────────────────────
 
   function rerenderTurn(turnId) {
     if (!isVisible()) {
@@ -672,293 +566,6 @@ export function createDebugPanel({ panelEl, bodyEl, clearBtnEl, sendInbound }) {
       bodyEl.insertBefore(node, bodyEl.firstChild);
     }
     turnNodes.set(turnId, node);
-  }
-
-  function renderTurn(turnId, turn) {
-    const section = document.createElement("section");
-    section.className = "debug-turn";
-    section.dataset.turnId = turnId;
-
-    const header = document.createElement("div");
-    header.className = "debug-turn-header";
-    header.title = "点击折叠 / 展开";
-    // 点击 header 切 ``.debug-turn-collapsed`` —— 整 section body 全收起，只剩 header。
-    // 实时 + 历史详情通吃；不抢内部 ``<details>``（候选 / 消息）的独立点击。
-    header.addEventListener("click", () => {
-      section.classList.toggle("debug-turn-collapsed");
-    });
-    const chevron = document.createElement("span");
-    chevron.className = "debug-turn-chevron";
-    chevron.textContent = "▾"; // ▾ 折叠态 CSS 旋转 -90deg 成 ▸
-    header.appendChild(chevron);
-    const ts = document.createElement("span");
-    ts.className = "debug-turn-ts";
-    ts.textContent = formatTs(turn.ts_ms);
-    header.appendChild(ts);
-
-    // 折叠态徽标：scoring_path（与 .debug-history-path 同款配色）。展开态由 CSS 隐藏。
-    if (turn.scoring_path) {
-      const path = document.createElement("span");
-      path.className = "debug-turn-path";
-      path.dataset.path = turn.scoring_path;
-      path.textContent = scoringPathLabel(turn.scoring_path);
-      header.appendChild(path);
-    }
-
-    const winnersWrap = document.createElement("div");
-    winnersWrap.className = "debug-turn-winners";
-    // 真实 winners = 本 turn 实际发言者（``MESSAGE_START`` 的 guest 集合）。
-    // envelope 不携带 orchestrator 的 ``winners`` 字段，``scores`` 包含全部候选（含
-    // 0.1 落选者），所以不能从 scores 反推；从 messages 推导既准确（哪怕只有 @mention
-    // 路由的确定性选手）又一致（pending 期 messages 为空，badge 自然不出现）。
-    const winnerNames = new Set();
-    for (const msg of turn.messages.values()) {
-      if (msg && msg.guest) winnerNames.add(msg.guest);
-    }
-    for (const w of winnerNames) {
-      winnersWrap.appendChild(makeBadge(w, "debug-winner-badge"));
-    }
-    // 折叠态用同一段 winner names 渲为纯文本（与 .debug-history-winners 同款）；
-    // 展开态由 CSS 隐藏 —— 让折叠/展开走同一份 DOM，避免重渲两套结构。
-    const winnersText = document.createElement("span");
-    winnersText.className = "debug-turn-winners-text";
-    winnersText.textContent = winnerNames.size > 0
-      ? [...winnerNames].join(", ")
-      : "（无人接话）";
-    winnersWrap.appendChild(winnersText);
-    header.appendChild(winnersWrap);
-
-    // 折叠态消息数（与 .debug-history-n 同款）。展开态由 CSS 隐藏。
-    if (turn.messages.size > 0) {
-      const n = document.createElement("span");
-      n.className = "debug-turn-n";
-      n.textContent = `${turn.messages.size} 条`;
-      header.appendChild(n);
-    }
-
-    section.appendChild(header);
-
-    // handoff turn 顶部提示条，与打分驱动的 turn 一眼区分。进 section body（非
-    // header）—— 折叠态随 body 一起收起。未知 handoff_* path 退回短标拼。
-    if (isHandoffPath(turn.scoring_path)) {
-      const note = document.createElement("div");
-      note.className = "debug-turn-handoff-note";
-      note.textContent = HANDOFF_NOTE_TEXT[turn.scoring_path]
-        || "由用户" + scoringPathLabel(turn.scoring_path);
-      section.appendChild(note);
-    }
-
-    // P9：兜底建桩的 turn（切回房间时已在后台进行）—— 顶部提示「明细不全」，
-    // 避免用户把空候选当成打分坏了。
-    if (turn.partial) {
-      const note = document.createElement("div");
-      note.className = "debug-turn-partial-note";
-      note.textContent =
-        "切回房间时这一轮已在后台进行 —— 切回前的打分候选 / 早前消息明细可能不全；"
-        + "完整取证见 debug/turns.jsonl（本轮结束后下次切房即补全）";
-      section.appendChild(note);
-    }
-
-    section.appendChild(renderCandidates(turn));
-
-    if (turn.messages.size > 0) {
-      section.appendChild(renderMessages(turn));
-    }
-
-    return section;
-  }
-
-  function renderCandidates(turn) {
-    const det = document.createElement("details");
-    det.className = "debug-section";
-    det.open = false;
-    const sum = document.createElement("summary");
-    sum.textContent = `候选（${turn.scores.length}）`;
-    det.appendChild(sum);
-
-    const ul = document.createElement("ul");
-    ul.className = "debug-candidates";
-    for (const s of turn.scores) {
-      ul.appendChild(renderCandidate(s, turn.scoring_prompts));
-    }
-    det.appendChild(ul);
-    return det;
-  }
-
-  function renderCandidate(score, scoringPrompts) {
-    const li = document.createElement("li");
-    li.className = "debug-candidate";
-    const row = document.createElement("div");
-    row.className = "debug-candidate-row";
-    const name = document.createElement("span");
-    name.className = "debug-candidate-name";
-    name.textContent = score.guest_name || "?";
-    row.appendChild(name);
-    const kind = document.createElement("span");
-    kind.className = "debug-candidate-kind";
-    kind.dataset.kind = score.kind || ScoreKind.SCORED;
-    kind.textContent = score.kind || ScoreKind.SCORED;
-    row.appendChild(kind);
-    const sc = document.createElement("span");
-    sc.className = "debug-candidate-score";
-    sc.textContent = formatScore(score.score);
-    row.appendChild(sc);
-    li.appendChild(row);
-
-    const det = document.createElement("details");
-    const sum = document.createElement("summary");
-    sum.textContent = "raw / prompt";
-    det.appendChild(sum);
-
-    if (score.raw) {
-      appendLabeledPre(det, "Scoring response (raw)", score.raw);
-    }
-    if (scoringPrompts === null) {
-      appendLabeledDisabled(det, "Scoring prompt");
-    } else if (scoringPrompts && scoringPrompts[score.guest_name]) {
-      appendLabeledPre(det, "Scoring prompt", scoringPrompts[score.guest_name]);
-    }
-    li.appendChild(det);
-    return li;
-  }
-
-  function renderMessages(turn) {
-    const det = document.createElement("details");
-    det.className = "debug-section";
-    det.open = true;
-    const sum = document.createElement("summary");
-    sum.textContent = `消息（${turn.messages.size}）`;
-    det.appendChild(sum);
-
-    const ul = document.createElement("ul");
-    ul.className = "debug-messages";
-    for (const [mid, msg] of turn.messages) {
-      ul.appendChild(renderMessage(mid, msg));
-    }
-    det.appendChild(ul);
-    return det;
-  }
-
-  function renderMessage(messageId, msg) {
-    const li = document.createElement("li");
-    li.className = "debug-message";
-    li.dataset.messageId = messageId;
-
-    const head = document.createElement("div");
-    head.className = "debug-message-head";
-    const guest = document.createElement("span");
-    guest.className = "debug-message-guest";
-    guest.textContent = msg.guest;
-    head.appendChild(guest);
-    const status = document.createElement("span");
-    status.className = "debug-status";
-    status.dataset.status = msg.status;
-    status.textContent = msg.status;
-    head.appendChild(status);
-    li.appendChild(head);
-
-    const det = document.createElement("details");
-    const sum = document.createElement("summary");
-    sum.textContent = "prompt / 工具 / 产物";
-    det.appendChild(sum);
-
-    if (msg.speak_prompt === null) {
-      appendLabeledDisabled(det, "Speak prompt");
-    } else {
-      appendLabeledPre(det, "Speak prompt", msg.speak_prompt);
-    }
-
-    if (msg.tool_calls.size > 0) {
-      const lab = document.createElement("div");
-      lab.className = "debug-pre-label";
-      lab.textContent = `工具调用（${msg.tool_calls.size}）`;
-      det.appendChild(lab);
-      const ul = document.createElement("ul");
-      ul.className = "debug-tools";
-      for (const [, t] of msg.tool_calls) {
-        ul.appendChild(renderTool(t));
-      }
-      det.appendChild(ul);
-    }
-
-    if (msg.artifact_paths.size > 0) {
-      const lab = document.createElement("div");
-      lab.className = "debug-pre-label";
-      lab.textContent = `产物（${msg.artifact_paths.size}）`;
-      det.appendChild(lab);
-      const ul = document.createElement("ul");
-      ul.className = "debug-artifacts";
-      for (const p of msg.artifact_paths) {
-        const ali = document.createElement("li");
-        ali.className = "debug-artifact";
-        ali.textContent = p;
-        ul.appendChild(ali);
-      }
-      det.appendChild(ul);
-    }
-
-    li.appendChild(det);
-    return li;
-  }
-
-  function renderTool(t) {
-    const li = document.createElement("li");
-    li.className = "debug-tool";
-
-    const head = document.createElement("div");
-    head.className = "debug-tool-head";
-
-    const name = document.createElement("span");
-    name.className = "debug-tool-name";
-    name.textContent = t.tool || "(unnamed)";
-    head.appendChild(name);
-
-    const src = document.createElement("span");
-    src.className = "debug-tool-source";
-    src.dataset.source = t.source;
-    src.textContent = t.source === "mcp" && t.mcp_server
-      ? `mcp:${t.mcp_server}`
-      : t.source;
-    head.appendChild(src);
-
-    if (typeof t.duration_ms === "number") {
-      const dur = document.createElement("span");
-      dur.className = "debug-tool-duration";
-      dur.textContent = `${t.duration_ms}ms`;
-      head.appendChild(dur);
-    }
-
-    const st = document.createElement("span");
-    st.className = "debug-tool-status";
-    st.dataset.status = t.status;
-    st.textContent = t.status;
-    head.appendChild(st);
-
-    li.appendChild(head);
-
-    if (t.args !== undefined && t.args !== null) {
-      // args 可能很大；rerenderTurn 每条 envelope 都会调一次本函数，缓在 entry 上
-      // 让同一条 tool_call 后续渲染（complete 帧 / 邻位变更）省 stringify 开销。
-      if (t._argsJson === undefined) t._argsJson = safeStringify(t.args);
-      const pre = makePre(t._argsJson);
-      pre.style.maxHeight = "120px";
-      li.appendChild(pre);
-    }
-    if (t.error) {
-      const err = document.createElement("div");
-      err.className = "debug-tool-error";
-      err.textContent = String(t.error);
-      li.appendChild(err);
-    }
-    return li;
-  }
-
-  function safeStringify(v) {
-    try {
-      return JSON.stringify(v, null, 2);
-    } catch {
-      return String(v);
-    }
   }
 
   return {
